Overhauled testing approach for text models
Before:
- models were random, non-deterministic in tests
- lots of complex regexs to try to find patterns that
  looked reasonably right
- still failed randomly
- hard to read
- low confidence in an actual response

Wherever a model picks a choice randomly from a list, we
now have a Choice abstraction. At runtime, a value
will be randomly chosen, but during tests it will
be explicit and deterministic.

Now tests will explicitly choose the choices before
the model does its thing. If a choice is missed and
the model attempts to pick one, the test will fail fast.

The result is a completely deterministic string of text
inline in the test.

Additionally, the tests make use of vitest's "snapshot"
feature, meaning if the logic changes, but it's still
acceptable, vitest can automatically update the assertions
with latest value.

Result: 100% deterministic and readable tests for random
generative models. Yay!