Fallback API key authenticator
This is an Authenticator implementation that can delegate
to multiple other Authenticators, allowing for new key types
to be introduced into the system, while maintaining backwards
compatibility with existing key types.

New keys will be generated with the primary authenticator.

Keys will be validated first with the primary, but if that fails
will iterate through secondaries until one either lets it through
or none remain and it fails.

Uses our good friend the Composite design pattern. If you know
Joe, you'll see this a lot :)